feat: Removendo os Mappers e atualizando a documentacao do projeto

+++ src/main/java/br/edu/fateczl/AvaliacaoDeliveryAV3/pedido/AtualizacaoPedido.java
@@ -11,7 +11,7 @@ public record AtualizacaoPedido(
         
         @NotNull(message = "Valor é obrigatório")
         @Positive(message = "Valor deve ser positivo")
-        Double valor,
+        Float valor,
         
         @NotBlank(message = "Cliente é obrigatório")
         String clienteCpf,
